# server/gatewaysvr/docs/swagger.yaml
definitions:
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  pb.GetQuestionVoPageRequest:
    properties:
      page:
        $ref: '#/definitions/pb.Page'
      question:
        $ref: '#/definitions/pb.QuestionInfo'
    type: object
  pb.JudgeCase:
    properties:
      inputs:
        description: 用例输入
        type: string
      outputs:
        description: 用例输出
        type: string
    type: object
  pb.JudgeConfig:
    properties:
      memoryLimit:
        description: 内存限制
        type: integer
      stackLimit:
        description: 堆栈限制
        type: integer
      timeLimit:
        description: 时间限制
        type: integer
    type: object
  pb.Page:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  pb.QuestionAddRequest:
    properties:
      answer:
        type: string
      content:
        type: string
      judgeCase:
        items:
          $ref: '#/definitions/pb.JudgeCase'
        type: array
      judgeConfig:
        $ref: '#/definitions/pb.JudgeConfig'
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  pb.QuestionInfo:
    properties:
      acceptedNum:
        description: 题目通过数
        type: integer
      answer:
        description: 题目答案
        type: string
      content:
        description: 内容
        type: string
      createTime:
        description: 创建时间
        type: string
      favourNum:
        description: 收藏数
        type: integer
      id:
        description: id
        type: integer
      isDelete:
        description: 是否删除
        type: integer
      judgeCase:
        description: 判题用例（json 数组）
        type: string
      judgeConfig:
        description: 判题配置（json 对象）
        type: string
      submitNum:
        description: 题目提交数
        type: integer
      tags:
        description: 标签列表（json 数组）
        type: string
      thumbNum:
        description: 点赞数
        type: integer
      title:
        description: 标题
        type: string
      updateTime:
        description: 更新时间
        type: string
      userId:
        description: 创建用户 id
        type: integer
    type: object
  pb.QuestionSubmitAddRequest:
    properties:
      code:
        type: string
      language:
        type: string
      questionId:
        type: integer
    type: object
  pb.QuestionSubmitQueryRequest:
    properties:
      language:
        type: string
      page:
        $ref: '#/definitions/pb.Page'
      questionId:
        type: integer
      status:
        type: integer
      userId:
        type: integer
    type: object
  pb.UserLoginRequest:
    properties:
      userName:
        description: 用户昵称
        type: string
      userPassword:
        description: 密码
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: swaggo测试API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: swaggo测试
  version: "1.0"
paths:
  /question/add:
    post:
      consumes:
      - application/json
      description: 添加问题
      parameters:
      - description: question
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.QuestionAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 添加问题
  /question/get/{id}:
    get:
      description: 根据id获取题目
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据id获取题目
  /question/list:
    post:
      consumes:
      - application/json
      description: 分页查询问题
      parameters:
      - description: QuestionVoPage
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.GetQuestionVoPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 分页查询问题
  /question/submit/do:
    post:
      consumes:
      - application/json
      description: 提交问题
      parameters:
      - description: QuestionSubmitAddRequest
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.QuestionSubmitAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 提交问题
  /question/submit/query:
    post:
      consumes:
      - application/json
      description: 查询问题提交信息
      parameters:
      - description: QuestionSubmit
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.QuestionSubmitQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 查询问题提交信息
  /test/{id}:
    get:
      description: 描述信息
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 测试接口
  /user/getLoginUser:
    get:
      description: 获取登录用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取登录用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/pb.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 用户登录
swagger: "2.0"